+++ run_integration_test.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+cd ./integration
+
+sh test.sh